funciones/bucket/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==2.6.0
